+++ internal/auth/auth.go
@@ -1,4 +1,4 @@
-package handlers
+package auth
 
 import (
 	"github.com/golang-jwt/jwt/v5"
